Struct pam::conv::PamConv

pub struct PamConv {
    // some fields omitted
}

PamConv acts as a channel for communicating with user.

Communication is mediated by the pam client (the application that invoked pam). Messages sent will be relayed to the user by the client, and response will be relayed back.

Methods

impl PamConv

fn send(&self, style: PamMessageStyle, msg: &str) -> PamResult<Option<String>>

Sends a message to the pam client.

This will typically result in the user seeing a message or a prompt. There are several message styles available:

PAM_PROMPT_ECHO_OFF

PAM_PROMPT_ECHO_ON

PAM_ERROR_MSG

PAM_TEXT_INFO

PAM_RADIO_TYPE

PAM_BINARY_PROMPT

Note that the user experience will depend on how the client implements these message styles - and not all applications implement all message styles.