Module pam::module

Functions for use in pam modules.

Enums

PamDataT

PamHandleT
Opaque type, used as a pointer when making pam API calls.

Traits

PamItem
Type-level mapping for safely retrieving values with get_item.

Functions

cleanup

get_data
Gets some value, identified by key, that has been set by the module previously.

get_item
Retrieves a value that has been set, possibly by the pam client. This is particularly useful for getting a PamConv reference.

get_user
Retrieves the name of the user who is authenticating or logging in.

set_data
Stores a value that can be retrieved later with get_data. The value lives as long as the current pam cycle.

Type Definitions

PamResult