Enum pam::module::PamHandleT

pub enum PamHandleT {}

Opaque type, used as a pointer when making pam API calls.

A module is invoked via an external function such as pam_sm_authenticate. Such a call provides a pam handle pointer. The same pointer should be given as an argument when making API calls.