Function pam::module::set_data

pub fn set_data<T>(pamh: &PamHandleT, key: &str, data: Box<T>) -> PamResult<()>

Stores a value that can be retrieved later with get_data. The value lives as long as the current pam cycle.

See pam_set_data in http://www.linux-pam.org/Linux-PAM-html/mwg-expected-by-module-item.html