Trait pam::module::PamItem

pub trait PamItem {
    fn item_type(_: Option<Self>) -> PamItemType;
}

Type-level mapping for safely retrieving values with get_item.

See pam_get_item in http://www.linux-pam.org/Linux-PAM-html/mwg-expected-by-module-item.html

Required Methods

fn item_type(_: Option<Self>) -> PamItemType

Maps a Rust type to a pam constant.

For example, the type PamConv maps to the constant PAM_CONV. The pam API contract specifies that when the API function pam_get_item is called with the constant PAM_CONV, it will return a value of type PamConv.

The argument will always be None. Its purpose is to provide a type label - the value is not important.